backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
requests
mercadopago
